backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-docx==0.8.11
pdfplumber==0.10.4
google-generativeai==0.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==11.0.0
